+++ test/test-lru.js
@@ -3,6 +3,8 @@ require('../src');
 const LRU = require('../src/datastore/lrucache');
 const UFCache = require('../src/datastore/ufcache');
 
+//===================//
+
 class MEM {
 	constructor () {
 		this._datastore = Object.create(null);
@@ -76,6 +78,72 @@ class TST {
 		}
 	}
 }
+
+class LRU2 {
+	constructor (limit=100) {
+		this._core = new Map();
+		this._cache = new Map();
+		this._secondary = new Map();
+		this._limit = isNaN(limit) ? 100 : limit;
+		this._gauge = Math.ceil(this._limit / 2);
+		this._length = 0;
+		this._count = 0;
+	}
+	set (k, v) {
+		const p = this._cache.get(k);
+		if (p === undefined) this._update(k, v);
+		else this._cache.set(k, v);
+	}
+	get (k) {
+		var v = this._core.get(k);
+		if (v !== undefined) return v;
+		v = this._cache.get(k);
+		if (v !== undefined) {
+			this._upgrade(k, v);
+			return v;
+		}
+		v = this._secondary.get(k);
+		if (v !== undefined) this._update(k, v);
+		return v;
+	}
+	del (k) {
+		var p = this._core.get(k);
+		if (p !== undefined) this._count --;
+		p = this._cache.get(k);
+		if (p !== undefined) this._length --;
+		this._core.delete(k);
+		this._cache.delete(k);
+		this._secondary.delete(k);
+	}
+	has (k) {
+		return this._cache.has(k) || this._secondary.has(k);
+	}
+	clear () {
+		this._cache = new Map();
+		this._secondary = new Map();
+		this._count = 0;
+		this._length = 0;
+	}
+	_update (k, v) {
+		this._length ++;
+		if (this._length >= this._limit) {
+			this._secondary = this._cache;
+			this._cache = new Map();
+			this._length = 0;
+		}
+		this._cache.set(k, v);
+	}
+	_upgrade (k, v) {
+		this._count ++;
+		if (this._count >= this._gauge) {
+			this._core = new Map();
+			this._core.set(k, v);
+			this._count = 0;
+		}
+	}
+}
+
+//===================//
 
 const M = {
 	_data: null,
@@ -166,11 +234,12 @@ const rndTester = (ds, label, limit=100, task=1000, loop=100) => {
 	start = Date.now();
 	for (let t = 0; t < loop; t ++) {
 		for (let i = 0; i < 10000; i ++) {
-			let j = Math.floor(Math.random() * task);
+			let t = Math.pick(rndTester.policy);
+			let j = t ? Math.floor(Math.random() * task) : Math.floor(Math.random() * limit);
 			let k = 'K-' + j, v = 'V: ' + j;
 			let vv = tester.get(k);
 			if (vv !== undefined && v !== vv) console.log(k, v, vv);
-			j = Math.floor(Math.random() * task);
+			j = t ? Math.floor(Math.random() * task) : Math.floor(Math.random() * limit);
 			k = 'K-' + j;
 			v = 'V: ' + j;
 			tester.set(k, v);
@@ -183,6 +252,7 @@ const rndTester = (ds, label, limit=100, task=1000, loop=100) => {
 
 	console.log('=============================');
 };
+rndTester.policy = 0.1;
 
 // var limit = 100, task = 100000000;
 // var limit = 100000000, task = 100;
@@ -199,6 +269,7 @@ var limit = 100, task = 1000000, loop = 100;
 // rndTester(MAP, 'Map', limit, task, loop);
 // rndTester(LRU, 'LRU', limit, task, loop);
 // rndTester(LRU.withDatastore, 'LRUWDS', limit, task, loop);
+rndTester(LRU2, 'LRU2', limit, task, loop);
+// rndTester(UFCache, 'UFCache', limit, task, loop);
+// rndTester(UFCache.withDatastore, 'UFCacheWDS', limit, task, loop);
 // rndTester(TST, 'TST', limit, task, loop);
-rndTester(UFCache, 'UFCache', limit, task, loop);
-rndTester(UFCache.withDatastore, 'UFCacheWDS', limit, task, loop);